rag-playground/requirements.txt
langchain
langchain_community
pypdf
faiss-cpu
pypdf
bs4
python-dotenv
openai
langchain-openai
gpt4all
streamlit
colorama
wikipedia
langchain-mistralai